backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File Upload
python-multipart==0.0.6
aiofiles==23.2.1

# CORS and HTTP
httpx==0.26.0

# Environment
python-dotenv==1.0.0
